+++ console/src/dataSources/services/mssql/index.tsx
@@ -6,6 +6,7 @@ import {
   Table,
   BaseTableColumn,
   SupportedFeaturesType,
+  FrequentlyUsedColumn,
   ViolationActions,
 } from '../../types';
 import { generateTableRowRequest, operators } from './utils';
@@ -118,15 +119,17 @@ export const supportedFeatures: SupportedFeaturesType = {
     },
     modify: {
       enabled: true,
-      readOnly: true,
       columns: {
         view: true,
-        edit: false,
+        edit: true,
+        graphqlFieldName: false,
+        frequentlyUsedColumns: false,
       },
+      readOnly: false,
       computedFields: false,
       primaryKeys: {
         view: true,
-        edit: false,
+        edit: true,
       },
       foreignKeys: {
         view: true,
@@ -134,7 +137,7 @@ export const supportedFeatures: SupportedFeaturesType = {
       },
       uniqueKeys: {
         view: true,
-        edit: false,
+        edit: true,
       },
       triggers: false,
       checkConstraints: {
@@ -318,11 +321,16 @@ FROM sys.objects as obj
                 WHEN t.is_user_defined = 1 THEN 'USER-DEFINED'
                 ELSE 'OTHER'
             END AS data_type,
-            t.name AS data_type_name
+            t.name AS data_type_name,
+            sch.name AS table_schema,
+            obj.name AS table_name,
+            ep.value AS comment,
+            ep.name AS extended_property_name_comment
         FROM
             sys.columns a
             LEFT JOIN sys.default_constraints ad ON (a.column_id = ad.parent_column_id AND a.object_id = ad.parent_object_id)
             JOIN sys.types t ON a.user_type_id = t.user_type_id
+            LEFT JOIN sys.extended_properties ep ON (ep.major_id = a.object_id AND ep.minor_id = a.column_id)
         WHERE a.column_id > 0 and a.object_id = obj.object_id
         FOR JSON path
 ) AS [isc](json) where obj.type_desc in ('USER_TABLE', 'VIEW') ${whereClause};`;
@@ -552,35 +560,131 @@ FROM sys.objects as obj
   getViewDefinitionSql: () => {
     return '';
   },
-  getDropColumnSql: () => {
+  getDropColumnSql: (
+    tableName: string,
+    schemaName: string,
+    columnName: string
+  ) => {
+    return `ALTER TABLE "${schemaName}"."${tableName}"
+    DROP COLUMN "${columnName}"`;
+  },
+  getAddColumnSql: (
+    tableName: string,
+    schemaName: string,
+    columnName: string,
+    columnType: string,
+    options?: {
+      nullable: boolean;
+      unique: boolean;
+      default: any;
+      sqlGenerator?: FrequentlyUsedColumn['dependentSQLGenerator'];
+    },
+    constraintName?: string
+  ) => {
+    let sql = `ALTER TABLE "${schemaName}"."${tableName}" ADD "${columnName}" ${columnType}`;
+    if (!options) {
+      return sql;
+    }
+    if (options.nullable) {
+      sql += ` NULL`;
+    } else {
+      sql += ` NOT NULL`;
+    }
+    if (options.unique) {
+      sql += ` UNIQUE`;
+    }
+    if (options.default) {
+      sql += ` CONSTRAINT "${constraintName}" DEFAULT '${options.default}' WITH VALUES`;
+    }
+    return sql;
+  },
+  getAddUniqueConstraintSql: (
+    tableName: string,
+    schemaName: string,
+    constraintName: string,
+    columns: string[]
+  ) => {
+    return `ALTER TABLE "${schemaName}"."${tableName}"
+    ADD CONSTRAINT "${constraintName}"
+    UNIQUE (${columns.join(',')})`;
+  },
+  getDropNotNullSql: (
+    tableName: string,
+    schemaName: string,
+    columnName: string,
+    columnType?: string
+  ) => {
+    return `ALTER TABLE "${schemaName}"."${tableName}" ALTER COLUMN "${columnName}"  ${columnType} NULL`;
+  },
+  getSetCommentSql: (
+    on: 'column' | 'table' | string,
+    tableName: string,
+    schemaName: string,
+    comment: string | null,
+    columnName?: string
+  ) => {
+    const dropCommonCommentStatement = `IF EXISTS (SELECT NULL FROM SYS.EXTENDED_PROPERTIES WHERE [major_id] = OBJECT_ID('${tableName}') AND [name] = N'${on}_comment_${schemaName}_${tableName}_${columnName}' AND [minor_id] = (SELECT [column_id] FROM SYS.COLUMNS WHERE [name] = '${columnName}' AND [object_id] = OBJECT_ID('${tableName}')))    
+        EXECUTE sp_dropextendedproperty   
+        @name = N'${on}_comment_${schemaName}_${tableName}_${columnName}',   
+        @level0type = N'SCHEMA', @level0name = '${schemaName}'
+    `;
+    const commonCommentStatement = `
+        exec sys.sp_addextendedproperty   
+        @name = N'${on}_comment_${schemaName}_${tableName}_${columnName}',   
+        @value = N'${comment}',   
+        @level0type = N'SCHEMA', @level0name = '${schemaName}'
+    `;
+    if (on === 'column') {
+      return `${dropCommonCommentStatement},@level1type = N'TABLE',  @level1name = '${tableName}',@level2type = N'COLUMN', @level2name = '${columnName}';
+      ${commonCommentStatement},@level1type = N'TABLE',  @level1name = '${tableName}',@level2type = N'COLUMN', @level2name = '${columnName}'`;
+    }
+    // FIXME: Comment on mssql table and function is not implemented yet.
     return '';
   },
-  getAddColumnSql: () => {
-    return '';
+  getSetColumnDefaultSql: (
+    tableName: string,
+    schemaName: string,
+    columnName: string,
+    defaultValue: any,
+    constraintName: string
+  ) => {
+    return `ALTER TABLE "${schemaName}"."${tableName}" DROP CONSTRAINT IF EXISTS "${constraintName}";
+    ALTER TABLE "${schemaName}"."${tableName}" ADD CONSTRAINT "${constraintName}" DEFAULT ${defaultValue} FOR "${columnName}"`;
   },
-  getAddUniqueConstraintSql: () => {
-    return '';
+  getSetNotNullSql: (
+    tableName: string,
+    schemaName: string,
+    columnName: string,
+    columnType: string
+  ) => {
+    return `ALTER TABLE "${schemaName}"."${tableName}" ALTER COLUMN "${columnName}"  ${columnType} NOT NULL`;
   },
-  getDropNotNullSql: () => {
-    return '';
+  getAlterColumnTypeSql: (
+    tableName: string,
+    schemaName: string,
+    columnName: string,
+    columnType: string,
+    wasNullable?: boolean
+  ) => {
+    return `ALTER TABLE "${schemaName}"."${tableName}" ALTER COLUMN "${columnName}" ${columnType} ${
+      !wasNullable ? `NOT NULL` : ``
+    }`;
   },
-  getSetCommentSql: () => {
-    return '';
+  getDropColumnDefaultSql: (
+    tableName: string,
+    schemaName: string,
+    columnName?: string,
+    constraintName?: string
+  ) => {
+    return `ALTER TABLE "${schemaName}"."${tableName}" DROP CONSTRAINT "${constraintName}"`;
   },
-  getSetColumnDefaultSql: () => {
-    return '';
-  },
-  getSetNotNullSql: () => {
-    return '';
-  },
-  getAlterColumnTypeSql: () => {
-    return '';
-  },
-  getDropColumnDefaultSql: () => {
-    return '';
-  },
-  getRenameColumnQuery: () => {
-    return '';
+  getRenameColumnQuery: (
+    tableName: string,
+    schemaName: string,
+    newName: string,
+    oldName: string
+  ) => {
+    return `sp_rename '[${schemaName}].[${tableName}].[${oldName}]', '${newName}', 'COLUMN'`;
   },
   fetchColumnCastsQuery: '',
   checkSchemaModification: () => {
@@ -589,8 +693,20 @@ FROM sys.objects as obj
   getCreateCheckConstraintSql: () => {
     return '';
   },
-  getCreatePkSql: () => {
-    return '';
+  getCreatePkSql: ({
+    schemaName,
+    tableName,
+    selectedPkColumns,
+    constraintName,
+  }: {
+    schemaName: string;
+    tableName: string;
+    selectedPkColumns: string[];
+    constraintName?: string;
+  }) => {
+    return `ALTER TABLE "${schemaName}"."${tableName}"
+    ADD CONSTRAINT "${constraintName}"
+    PRIMARY KEY (${selectedPkColumns.map(pkc => `"${pkc}"`).join(',')})`;
   },
   getFunctionDefinitionSql: null,
   primaryKeysInfoSql: ({ schemas }) => {